move methods into intermediate class
