+++ app/lib/submission_builder/ty2022/states/az/individual_return.rb
@@ -3,7 +3,7 @@ module SubmissionBuilder
   module Ty2022
     module States
       module Az
-        class IndividualReturn < SubmissionBuilder::Document
+        class IndividualReturn < StateReturn
           include DependentRelationshipTable
 
           FILING_STATUS_OPTIONS = {
